Replace the internal custom logger with Log4j standard loggers
The primary work here is to remove the internal custom logging library
and replace it with a standard Log4j setup. This lets the user define
their own logging settings in a standard way that doesn't require a
bunch of custom settings in the AgentConfiguration class.

Unfortunately there are a few other fixes in here that we've squashed
down - I've tried to list them here:

* Fix a few bugs in the Ubuntu init script
* Dynamically find the 'service' command path in the babysit script
* Leverate the actual init scripts for auto-restarting the agent if necessary.
* A few more Ubuntu init script bug fixes.
* Package the app up into amazon-kinesis-agent-<tag>.tar.gz files.
* Bugfix: /etc/default, not /etc/defaults on Ubuntu
* Bugfix on the babysit script: --ppid doesn't work everywhere.
* Remove custom logger