Buid and ship a Fat JAR with Dependencies
When we build out the tarball, we no longer need to use
the setup script at all - instead, we do a mvn dependency download
into a local .m2/repository directory, and reference that
directory later in the setup script. This lets us keep building
the pre-cached-download tarball as-is.

Ultimately, the tarball becomes less useful though because
we are now also building a fully packaged jar file with all
of the dependencies baked in.

Additionally, none of the builds were being run on-commit because of
some problems in the overall file-structure in the repo. I cleaned those
up, patched up the POM file and more.